updated and implemented in API removeSurgery and tested, updateSurgery with testing, implement in frontend removeSurgery and improved navBar #455

+++ staff/jordi-relano/innospine/api/logic/removeSurgery.ts
@@ -2,11 +2,11 @@
  import { validate, errors } from 'com'
  import { Surgery, User } from '../data/index.ts'
  
- 
  const { SystemError, NotFoundError } = errors
  
  function removeSurgery(surgeryId: string, userId: string): Promise<void> {
      validate.text(surgeryId, 'surgeryId', true)
+     validate.text(userId, 'surgeryId', true)
  
      return Surgery.findOne({ _id: surgeryId })
          .catch(error => { throw new SystemError(error.message) })
@@ -14,18 +14,18 @@
              if (!surgery)
                  throw new NotFoundError('surgery not found')            
  
-             User.findById(userId)
+            return User.findById(userId)
                  .catch(error => { throw new SystemError(error.message) })
                  .then(user => {
                      if (!user)
                          throw new NotFoundError('user does not exist')
  
-                     if (surgery.author !== user._id)
-                         throw new NotFoundError('surgery does not belong to user')
+                     if (surgery.author._id.toString() !== user._id.toString())
+                        throw new NotFoundError('surgery does not belong to user')
+
+                    return Surgery.deleteOne({ _id: surgeryId, author: userId })
+                        .catch(error => { throw new SystemError(error.message) })
                  })
-             
-             return Surgery.deleteOne({ _id: surgeryId })
-                 .catch(error => { throw new SystemError(error.message) })
          })
  }
  export default removeSurgery
